Make the batch as large as possible
The inbox contains items only up to the next watermark. If WMs are frequent,
this severely limits the batch size. This commit changes the behavior of
AsyncTransformUsingServiceBatchedP to collect items until max batch size is
reached or tryProcess() is called twice (so that we know there are no more
items immediately available).

Fixes #1983